refactoring: extract pagination utils
